+++ NEWS.md
@@ -39,6 +39,7 @@
 * `calc_ne()` and `run_colony()` actually cleanup now if asked on Windows.
 * Fixed the `verbose` argument on `run_colony()` to actually work on Windows.
 * Fixed a bug trying to merge new and old data with the same columns in different orders could cause NAs to be filled by the wrong numbers. Internal only problem that only effected the new `calc_abba_baba()` function.
+* Fixed a bug with `calc_ne()` where pop names were not being properly handled and het/coan method results were not being returned by `get.snpR.stats()`.
 
 # snpR 1.2.1